agentcompose: remove session bridge wrappers
